Added DotVVM project news
I am not sure whether the blog post is appropriate for the newsletter - we are looking for volunteers to help us create content about DotVVM and we are able to pay for that, and I think the message is relevant for the subscribers. But technically it is a job offer and I am not sure whether you want such content in the newsletter, so please reject if you think it doesn't fit here.