Avoid using US-centric date format
Particularly when it doesn't contain a year, but the year is the same as the day-of-month.

I'd generally suggest we either use words ("February 20th") or ISO-8601 dates ("2020-02-20"). My experience with working in design docs with US colleagues is that either of these formats will cause significantly less discombobulation (at least for me).